Pull out components into files
